Application divided according to bussiness logic into programmers and accounts apps.
