+++ knownProgrammer/accounts/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
 
-from .models import CustomUser, ProgrammerProfile
+from .models import CustomUser
 
 admin.site.register(CustomUser)
-admin.site.register(ProgrammerProfile)
