+++ knownProgrammer/envtemplate.txt
@@ -9,6 +9,7 @@ DB_HOST =
 CLOUD_NAME =
 API_KEY =
 API_SECRET =
+DEFAULT_AVATAR =
 
 EMAIL_BACKEND = ''
 EMAIL_HOST = ''
