+++ knownProgrammer/programmers/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from .models import ProgrammerProfile
+
+admin.site.register(ProgrammerProfile)
