+++ knownProgrammer/programmers/templatetags/stars.py
